fixup(m2): adapt to M9 csrf_cookie_name relocation
main moved csrf_cookie_name from tests/e2e/conftest.py to
tests/e2e/helpers.py. Update the catalog/runtime E2E imports to
match.
