+++ backend/tests/e2e/test_mcp_catalog_routes.py
@@ -44,7 +44,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from cubebox.models import MCPCatalogConnector, MCPServer
 from cubebox.repositories.mcp_catalog import MCPCatalogConnectorRepository
-from tests.e2e.conftest import _csrf_cookie_name
+from tests.e2e.helpers import csrf_cookie_name as _csrf_cookie_name
 
 # ---------------------------------------------------------------------------
 # Catalog seeding helpers
